requirements.txt
streamlit
python-dotenv
Pillow
langchain-core
langchain-google-genai
